requirements.txt
flask==2.3.3
groq==0.9.0
python-dotenv==1.0.0
httpx==0.24.1
